re-sort python code
