+++ python/convert_syslog_priority.py
@@ -5,157 +5,17 @@ Convert given syslog priority value into facility and severity values.
 """
 
 __author__: str = "Daniel Kuß"
-__version__: str = "1.1.0"
+__version__: str = "1.2.0"
 __src__: str = "https://github.com/Cthullu/syslog_priority_converter"
 __status__: str = "Production"
 
 
 # Standard imports
-import argparse
-import logging
 from sys import exit as sys_exit
-from typing import Optional
 
 # Local imports
-from syslog_converter import facility
-from syslog_converter import severity
-
-
-LOGLEVELS = {
-    'CRITICAL': logging.CRITICAL,
-    'ERROR': logging.ERROR,
-    'WARNING': logging.WARNING,
-    'INFO': logging.INFO,
-    'DEBUG': logging.DEBUG,
-    'NOTSET': logging.NOTSET
-}
-
-
-def get_parser() -> argparse.ArgumentParser:
-    """
-    Returns an ArgumentParser for the convert_syslog_priority.py script.
-
-    Args:
-        None
-
-    Returns:
-        argparse.ArgumentParser: Configured argument parser.
-    """
-    parser = argparse.ArgumentParser(
-        prog = "convert_syslog_priority",
-        description="Convert syslog priority value to facility and severity level.",
-    )
-
-    parser.add_argument(
-        "-l", "--loglevel",
-        choices = LOGLEVELS,
-        default = "WARNING",
-        dest = "loglevel",
-        help = "Select logging level.",
-        metavar = "<loglevel>",
-        required = False,
-        type = str.upper,
-    )
-
-    parser.add_argument(
-        "-v", "--version",
-        action = "version",
-        version = f'%(prog)s {__version__}'
-    )
-
-    parser.add_argument(
-        dest = "priority",
-        type = int,
-        metavar = "<priority>",
-        help = "Priority to convert into facility and severity level.",
-    )
-
-    return parser.parse_args()
-
-
-def setup_logger(loglevel: Optional[int] = LOGLEVELS["WARNING"]) -> logging.Logger:
-    """
-    Set up the logging configuration for the script.
-
-    Args:
-        loglevel (Optional[int]): The logging level to set. Defaults to WARNING.
-
-    Returns:
-        logging.Logger: Configured logger instance.
-    """
-    logger = logging.getLogger()
-    logger.setLevel(loglevel)
-
-    # Create a console handler with the specified log level
-    console_handler = logging.StreamHandler()
-    console_handler.setLevel(loglevel)
-    formatter = logging.Formatter('%(asctime)s - %(levelname) -8s - %(message)s')
-    console_handler.setFormatter(formatter)
-    logger.addHandler(console_handler)
-
-    return logger
-
-
-def extract_values(priority: int, logger: Optional[logging.Logger] = None) -> dict:
-    """
-    Extract the facility and severity value from a provided priority.
-    Also adds the facility and severity name.
-
-    Args:
-        priority (int): Syslog priority value.
-        logger (Optional[Logging.logger]): Logger instance for logging debug messages.
-
-    Returns:
-        dict: Dictionary containing facility value, facility keyword, severity value,
-            and severity keyword.
-    """
-    ret_val = {}
-
-    logger.debug("Get facility value from provided priority '%s'.", priority)
-    try:
-        ret_val["facility_value"] = facility.get_facility_value(priority)
-    except TypeError as exc:
-        logger.error("Invalid type for priority value: %s", exc)
-        raise TypeError("Priority must be an integer.") from exc
-    except ValueError as exc:
-        logger.error("Invalid priority value: %s", exc)
-        raise ValueError("Priority must be between 0 and 191 inclusive.") from exc
-
-    logger.debug("Get facility keyword for value '%s'.",ret_val["facility_value"])
-    try:
-        ret_val["failicty_keyword"] = facility.get_facility_keyword(
-            ret_val["facility_value"]
-        )
-    except KeyError as exc:
-        logger.error("Invalid facility value: %s", exc)
-        raise KeyError("Facility must be between 0 and 23 inclusive.") from exc
-    except TypeError as exc:
-        logger.error("Invalid type for facility value: %s", exc)
-        raise TypeError("Facility must be an integer.") from exc
-
-    logger.debug("Get severity level from provided priority '%s'.", priority)
-    try:
-        ret_val["severity_value"] = severity.get_severity_value(priority)
-    except TypeError as exc:
-        logger.error("Invalid type for priority value: %s", exc)
-        raise TypeError("Priority must be an integer.") from exc
-    except ValueError as exc:
-        logger.error("Invalid priority value: %s", exc)
-        raise ValueError("Priority must be between 0 and 191 inclusive.") from exc
-
-    logger.debug("Get severity keyword for value '%s'.", ret_val["severity_value"])
-    try:
-        ret_val["severity_keyword"] = severity.get_severity_keyword(
-            ret_val["severity_value"]
-        )
-    except KeyError as exc:
-        logger.error("Invalid severity value: %s", exc)
-        raise KeyError("Severity must be between 0 and 7 inclusive.") from exc
-    except TypeError as exc:
-        logger.error("Invalid type for severity value: %s", exc)
-        raise TypeError("Severity must be an integer.") from exc
-
-    return ret_val
+from syslog_converter import converter
+from syslog_converter import utils
 
 
 def main() -> int:
@@ -168,26 +28,33 @@ def main() -> int:
     Returns:
         int: Exit code, 0 on success, 1 on error.
     """
-    logger = setup_logger("INFO")
+    logger = utils.setup_logger("INFO")
 
     logger.info("Starting syslog priority conversion script.")
 
-    cli_args = get_parser()
+    cli_args = utils.get_cli_args(__version__)
     if cli_args.loglevel != "INFO":
-        logger = setup_logger(cli_args.loglevel)
+        logger = utils.setup_logger(cli_args.loglevel)
 
-    logger.debug("Call subfunction to extract values from priority.")
+    logger.debug("Creating converter instance with priority '%d'.", cli_args.priority)
     try:
-        extracted_values = extract_values(cli_args.priority, logger)
-    except (TypeError, ValueError, KeyError) as exc:
-        logger.error("Error extracting values: %s", exc)
-        print(f"Error: {exc}")
+        priority_converter = converter.Converter(cli_args.priority)
+    except ValueError as e:
+        logger.error(
+            "Received value error while creating converter object with priority '%d': %s",
+            cli_args.priority, e
+        )
         return 1
 
-    print(f"Facility: {extracted_values['facility_value']} "
-          f"({extracted_values['failicty_keyword']})")
-    print(f"Severity: {extracted_values['severity_value']} "
-          f"({extracted_values['severity_keyword']})")
+    logger.debug("Getting facility and severity values from converter object.")
+    facility = priority_converter.facility
+    facility_level = priority_converter.facility_level
+    severity = priority_converter.severity
+    severity_level = priority_converter.severity_level
+
+    logger.debug("Printing values.")
+    print(f"Syslog severity: {severity} {facility_level}")
+    print(f"Syslog facility: {facility} {severity_level}")
 
     logger.info("Syslog priority conversion completed successfully.")
     logger.debug("Exiting with code 0.")
